Add dynamic annotated bibliography
This proof-of-concept uses Zotero to store the items (and its API to
retrieve them) and citeproc.js to render IEEE citations. Annotations are
stored in the Zotero `Extra` field, prefixed with 'Annotation:'. The
table always pulls the latest data from Zotero, so we only need to make
updates in one place.